+++ cmd/lineage/main.go
@@ -8,7 +8,7 @@ import (
 )
 
 func main() {
-	if err := cli.Execute(context.Background(), os.Args[1:], os.Stdout, os.Stderr); err != nil {
+	if err := cli.Execute(context.Background(), os.Args[1:], os.Stdin, os.Stdout, os.Stderr); err != nil {
 		os.Exit(1)
 	}
 }